Refactor: Products api 관리방법 및 방법 변경으로 인한 import 수정
